Design for achievments #49
I also realised that the content after a section element is floating in the section. It had no effect on the content but was visible in the inspector. In order to prevent that I spamed a clearfix-element after every section because we can not fix the problem with an after-clearfix.