Removed background on team navbar-title #69
Background is now removed when you hover... Also the cursor did not switched to pointer when the links in the subpages got hovered... Its fixed now. teampage-title shows the text-selector cursor since there is nothing to click. It is also full white since it is selected.

Hovering is only fixed on dektop right now...