Fixed typo in the new shadow variables
I still don't know if this are the correct numbers for using material shadows in css... Goggle released this numbers only quite recently and they are used for sketch, but their meaning is the same. I am also not sure if it is correct to use 3 shadows for one since our shadows before only used 2 but Google is giving us 3 shadows... I have to wait and see if some nice online articles about this pop up in the future... At least the shadows are looking good in my opinion.